1. RESTFul. Описание архитектуры [#6884 #408988]

+++ src/main/java/ru/job4j/service/SpringPersonService.java
@@ -21,12 +21,25 @@ public class SpringPersonService {
         return personRepository.findById(id);
     }
 
-    public Person save(Person person) {
+    public Person create(Person person) {
         return personRepository.save(person);
     }
 
+    public boolean save(Person person) {
+        var res = personRepository.findById(person.getId());
+        if (res.isPresent()) {
+            personRepository.save(person);
+            return true;
+        }
+        return false;
+    }
+
     public boolean delete(Person person) {
-        personRepository.delete(person);
-        return !personRepository.existsById(person.getId());
+        var res = personRepository.findById(person.getId());
+        if (res.isPresent()) {
+            personRepository.delete(person);
+            return true;
+        }
+        return false;
     }
 }
